requirements.txt
Package           Version   
----------------- ----------
alembic           1.5.8     
appdirs           1.4.3     
attrs             20.3.0    
CacheControl      0.12.6    
certifi           2019.11.28
chardet           3.0.4     
click             7.1.2     
colorama          0.4.3     
ConfigArgParse    1.4       
contextlib2       0.6.0     
distlib           0.3.0     
distro            1.4.0     
fastapi           0.63.0    
Flask             1.1.2     
Flask-BasicAuth   0.2.0     
gevent            21.1.2    
geventhttpclient  1.4.4     
greenlet          1.0.0     
h11               0.12.0    
html5lib          1.0.1     
idna              2.8       
iniconfig         1.1.1     
ipaddr            2.2.0     
itsdangerous      1.1.0     
Jinja2            2.11.3    
lockfile          0.12.2    
locust            1.4.3     
Mako              1.1.4     
MarkupSafe        1.1.1     
msgpack           0.6.2     
packaging         20.3      
pep517            0.8.2     
pip               20.0.2    
pkg-resources     0.0.0     
pluggy            0.13.1    
progress          1.5       
psutil            5.8.0     
psycopg2-binary   2.8.6     
py                1.10.0    
pydantic          1.8.1     
pyparsing         2.4.6     
pytest            6.2.2     
python-dateutil   2.8.1     
python-dotenv     0.15.0    
python-editor     1.0.4     
pytoml            0.1.21    
pyzmq             22.0.3    
requests          2.22.0    
retrying          1.3.3     
setuptools        44.0.0    
six               1.14.0    
SQLAlchemy        1.3.23    
SQLAlchemy-Utils  0.36.8    
starlette         0.13.6    
toml              0.10.2    
typing-extensions 3.7.4.3   
urllib3           1.25.8    
uvicorn           0.13.4    
webencodings      0.5.1     
Werkzeug          1.0.1     
wheel             0.34.2    
zope.event        4.5.0     
zope.interface    5.3.0     
